transform LegFortescue class to Interface
