+++ iidm/iidm-extensions/src/main/java/com/powsybl/iidm/network/extensions/ThreeWindingsTransformerFortescue.java
@@ -10,8 +10,6 @@ package com.powsybl.iidm.network.extensions;
 import com.powsybl.commons.extensions.Extension;
 import com.powsybl.iidm.network.ThreeWindingsTransformer;
 
-import java.util.Objects;
-
 /**
  * @author Jean-Baptiste Heyberger {@literal <jbheyberger at gmail.com>}
  * @author Geoffroy Jamgotchian {@literal <geoffroy.jamgotchian at rte-france.com>}
@@ -19,94 +17,6 @@ import java.util.Objects;
 public interface ThreeWindingsTransformerFortescue extends Extension<ThreeWindingsTransformer> {
 
     String NAME = "threeWindingsTransformerFortescue";
-
-    class LegFortescue {
-
-        private double rz;
-        private double xz;
-        private boolean freeFluxes;
-        private WindingConnectionType connectionType;
-        private double groundingR;
-        private double groundingX;
-
-        public LegFortescue(double rz, double xz, boolean freeFluxes, WindingConnectionType connectionType,
-                            double groundingR, double groundingX) {
-            this.rz = rz;
-            this.xz = xz;
-            this.freeFluxes = freeFluxes;
-            this.connectionType = Objects.requireNonNull(connectionType);
-            this.groundingR = groundingR;
-            this.groundingX = groundingX;
-        }
-
-        /**
-         * Free fluxes set to true means that the magnetizing impedance Zm is infinite, i.e. fluxes are free.
-         */
-        public boolean isFreeFluxes() {
-            return freeFluxes;
-        }
-
-        public void setFreeFluxes(boolean freeFluxes) {
-            this.freeFluxes = freeFluxes;
-        }
-
-        /**
-         * The zero sequence resistance of the leg.
-         */
-        public double getRz() {
-            return rz;
-        }
-
-        public void setRz(double rz) {
-            this.rz = rz;
-        }
-
-        /**
-         * The zero sequence reactance of the leg.
-         */
-        public double getXz() {
-            return xz;
-        }
-
-        public void setXz(double xz) {
-            this.xz = xz;
-        }
-
-        /**
-         * Get the winding connection type of the leg, see {@link WindingConnectionType}).
-         */
-        public WindingConnectionType getConnectionType() {
-            return connectionType;
-        }
-
-        public void setConnectionType(WindingConnectionType connectionType) {
-            this.connectionType = Objects.requireNonNull(connectionType);
-        }
-
-        /**
-         * If the leg is earthed, depending on {@link WindingConnectionType}, it represents the
-         * resistance part of the impedance to ground.
-         */
-        public double getGroundingR() {
-            return groundingR;
-        }
-
-        public void setGroundingR(double groundingR) {
-            this.groundingR = groundingR;
-        }
-
-        /**
-         * If the leg is earthed, depending on {@link WindingConnectionType}, it represents the
-         * reactance part of the impedance to ground.
-         */
-        public double getGroundingX() {
-            return groundingX;
-        }
-
-        public void setGroundingX(double groundingX) {
-            this.groundingX = groundingX;
-        }
-    }
 
     @Override
     default String getName() {
